css-align-3/content-distribution/place-content-shorthand-004.html

<!DOCTYPE html>
<title>CSS Box Alignment: place-content shorthand - invalid values</title>
<link rel="author" title="Javier Fernandez" href="mailto:jfernandez@igalia.com" />
<link rel="help" href="http://www.w3.org/TR/css3-align/#propdef-place-content" />
<meta name="flags" content="invalid" />
<meta name="assert" content="Check that place-content's invalid values are properly detected at parsing time." />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
    #test {
        align-content: start;
        justify-content: end;
    }
</style>
<div id="log"></div>
<div id="test"></div>
<script>
    function checkInvalidValues(value) {
        var div = document.getElementById("test");
        div.setAttribute("style", "place-content: " + value);
        var style = getComputedStyle(div);
        assert_equals(style.getPropertyValue("align-content"),
            "start", "align-content computed value");
        assert_equals(style.getPropertyValue("justify-content"),
            "end", "justify-content computed value");
    }

    test(function() {
        checkInvalidValues("center safe")
        checkInvalidValues("true center")
    }, "Verify overflow keywords are invalid");

    test(function() {
        checkInvalidValues("center space-between start")
    }, "Verify fallback values are invalid");

    test(function() {
        checkInvalidValues("10px left")
        checkInvalidValues("right 10%")
    }, "Verify numeric values are invalid");

    test(function() {
       checkInvalidValues("auto right")
       checkInvalidValues("auto auto")
    }, "Verify 'auto' values are invalid");

    test(function() {
        checkInvalidValues("")
    }, "Verify empty declaration is invalid");
</script>
